fix: distinguish extension search sources
